requirements.txt
BeautifulSoup4==4.8.1
werkzeug==1.0.1
python-dateutil==2.8.1
requests==2.22.0
mysql-connector-python==8.0.19
pyyaml==5.3
python-telegram-bot==12.3.0
python-dateutil==2.8.1
pytest==6.2.5

git+https://github.com/pckmsolutions/pmgsiteinfra.git@v0.1.8
